Rework data structure
